pyproject.toml
[project]
name = "py-mail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=4.0.1",
    "fastapi>=0.115.12",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
